feat: 결제 시 SALES 테이블에 값 추가

+++ backend/store/src/main/java/com/ssafy/api/request/sales/SalesUpdatePutRequest.java
@@ -13,5 +13,5 @@ public class SalesUpdatePutRequest {
 	@ApiParam(value = "가게 아이디", required = true)
 	String storeId;
 	@ApiParam(value = "판매 금액", example = "3000", required = true)
-	Integer sum;
+	Long sum;
 }
